+++ algorithms/__init__.py
@@ -10,6 +10,6 @@ from algorithms import dijkstra_waypoint
 # Limit-cycle navigation
 from algorithms import limit_cycle
 
-from algorithms.univector_field import UnivectorField
+from algorithms import univector_field
 
 from algorithms.RRT import rrt
